#1080 move sidebar to private app
