+++ apps/private/src/components/Calendar/CalendarLayout.tsx
@@ -14,12 +14,10 @@ import CalendarController, {
 } from '@common/components/Calendar/CalendarController'
 import { CALENDAR_VIEWS } from '@common/components/Calendar/utils/constants'
 import { setView } from '@common/features/Settings/SettingsSlice'
-import Sidebar from '@common/components/Calendar/Sidebar/SideBar'
+import Sidebar from './Sidebar/SideBar'
 import TempSearchDialog from '@common/components/Calendar/TempSearchDialog'
 import { setIsMobileSearchOpen } from '@common/features/Calendars/CalendarSlice'
 import { useManageCalendarSelection } from './hooks/useManageCalendarSelection'
-
-import { CreateAppointmentModal } from '../../features/booking/CreateAppointmentModal'
 
 export default function CalendarLayout(): JSX.Element {
   const calendarRef = useRef<CalendarApi | null>(null)
@@ -56,22 +54,6 @@ export default function CalendarLayout(): JSX.Element {
 
     setViewMode()
   }, [isTablet, isMobile])
-
-  const [isCreateAppointmentModalOpen, setIsCreateAppointmentModalOpen] =
-    useState(false)
-
-  useEffect(() => {
-    const handleOpenModal = (): void => {
-      setIsCreateAppointmentModalOpen(true)
-    }
-    window.addEventListener('open-create-appointment-modal', handleOpenModal)
-    return () => {
-      window.removeEventListener(
-        'open-create-appointment-modal',
-        handleOpenModal
-      )
-    }
-  }, [])
 
   const isInIframe = useMemo(() => new CozyBridge().isInIframe(), [])
 
@@ -206,10 +188,6 @@ export default function CalendarLayout(): JSX.Element {
         <SettingsPage menubarProps={menubarProps} isInIframe={isInIframe} />
       )}
       <ErrorSnackbar error={error} type="calendar" />
-      <CreateAppointmentModal
-        open={isCreateAppointmentModalOpen}
-        onClose={() => setIsCreateAppointmentModalOpen(false)}
-      />
     </div>
   )
 }
