+++ apps/private/src/components/Calendar/Sidebar/DesktopSidebar.tsx
