+++ apps/private/src/components/Calendar/Sidebar/TabletSidebar.tsx
