+++ apps/private/src/components/Calendar/Sidebar/ViewSwitcher.tsx
